Move move ui/data to testing/data
