+++ frontend/src/ui/Select.stories.tsx
@@ -1,8 +1,8 @@
 // also exported from '@storybook/react' if you can deal with breaking changes in 6.1
 import { Meta, Story } from '@storybook/react/types-6-0';
 import React from 'react';
-import occupations from './data/occupations';
-import states from './data/states';
+import occupations from '../testing/data/occupations';
+import states from '../testing/data/states';
 import {
   OccupationSelect,
   OccupationSelectProps,
