+++ frontend/src/ui/TransitionTable.stories.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 // also exported from '@storybook/react' if you can deal with breaking changes in 6.1
 import { Story, Meta } from '@storybook/react/types-6-0';
 import TransitionTable, { TransitionTableProps } from './TransitionTable';
-import transitionData from './data/transitionData';
+import transitionData from '../testing/data/transitionData';
 
 export default {
   title: 'JobHopper/TransitionTable',
